+++ zarrs_metadata/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Add `UnsupportedAdditionalFieldError::new`
+
+### Fixed
+- Make `AdditionalField` public and permit any JSON type (not just objects)
+
 ## [0.3.3] - 2025-02-06
 
 ### Fixed